Move `useSelectedTables` one level up
